backend/requirements.txt
streamlit>=1.35.0
openai>=1.30.0
python-dotenv>=1.0.0
plotly>=5.20.0
pandas>=2.1.0
